Remove remaining calls to 'Router' singleton
Early versions of Troposphere used a singleton of `<Router/>` in order
to "transitionTo" new locations following the completion of an action.

With the new versions of react-router (and modified APIs), this is no
longer a feasible approach. When navigating from a non-RouteComponent,
you have have two choices in v2.8.1 [0]:

- using `push` on `this.context.router`
- use the `browserHistory` singleton

So code that used to change location the Troposphere defined "Router",
like so:

```
Router.getInstance().transitionTo("projects");
```

Becomes:
```
import { browserHistory } from 'react-router'
// ...
// ...

    browserHistory.push(`/projects`);

```

This scenario is in play any time we are looking to change location &
we are not within a RouteComponent [1].

[0] https://github.com/ReactTraining/react-router/blob/v2.8.1/upgrade-guides/v2.0.0.md#navigating-inside-deeply-nested-components
[1] https://github.com/ReactTraining/react-router/blob/v2.8.1/docs/Glossary.md#routecomponent