Include disable attribute in list items
If an action-link has `isDisabled: true`, but is not opened in a new
window, then the action-link is always going to appear as _enabled_.

This is the case for "Web Desktop" action-link that is not doing a
form POST and not an "open in new tab" via `href`. So the handling
of it meant that a suspended or stopped instance might have an enabled
action-link.

Note, any action-link within `isDisable` will be falsy, and thus appear
as enabled - which would be reasonable.

See ATMO-1510 [0]

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1510