Define core implementation for instance action
An "instance action" has many commonalities that are previously been
repeated within the "deleted" action definitions. This provides a
common (almost Template Method) implementation for handling an action
for a single instance.

This `instanceActionImpl` method is not exported as we would like to
support applying these actions to a collection of instances, not just
one.

If a single instance needs an action, it can be wrapped in a collection
and passed on to `applyInstanceAction` with the same result.

We may considering renaming the exported function if it is determined
that "apply" is a confusing term to use here.