Ensure retrieving HREF from model is safe
Here, we mean "safe" in that an attempt to `.get("href")` will not
result in an error.

This is the errors noted in other tickets [0]. We see an error relates
to potentially two links to help documents [1]. In this commit, we adopt
the same "empty" document fragment [2] link we do in 4389a3c9a (and, it
still could pop the view up to the top of the viewport. But, it will not
cause an unhandled error.

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1752
[1] https://pods.iplantcollaborative.org/jira/browse/ATMO-1753
[2] https://www.w3.org/TR/html5/browsers.html#scroll-to-fragid