Allow mixins w/ JSX to ignore displayName warning
These two files definition partial/mixin components that do not benefit
from identifying themselves with the `displayName` property. This
property is used by the React DevTools to show the component name when
inspecting. Since the partial/mixin may appear within multiple elements
it is not likely to be as useful as just identifying the component by
name (the component making use of the mixin).